Serializing data

+++ src/mmr/include/settings.h
@@ -6,11 +6,11 @@ namespace param {
 static const int TARGET_VALUE = 5000;
 static const int BIN_SIZE = 10;
 //
-// !! These max values are only valid for normalized meshes !!
+// !! These max m_values are only valid for normalized meshes !!
 //
 #define PI 3.14159265
 // Max angle
-static const float RAD_TO_DEG = 180 / PI;
+static const float RAD_TO_DEG = static_cast<float>(180 / PI);
 static const float A3_MAX_VALUE = 180;
 // Max distance from vertex to centroid
 static const float D1_MAX_VALUE = 1;
